Sync to current parent

+++ core/tag_api.php
@@ -55,6 +55,8 @@ require_api( 'string_api.php' );
 require_api( 'user_api.php' );
 require_api( 'utility_api.php' );
 
+use Mantis\Exceptions\ClientException;
+
 # cache the tag definitions, indexed by tag id
 # tag ids that don't exist are stored as 'false', to avoid repeated searches
 $g_cache_tags = array();
@@ -73,20 +75,20 @@ $g_cache_bug_tags = array();
 function tag_cache_rows( array $p_tag_ids ) {
 	global $g_cache_tags;
 
-	$t_ids_to_seach = array();
+	$t_ids_to_search = array();
 	foreach( $p_tag_ids as $t_id ) {
 		if( !isset( $g_cache_tags[(int)$t_id]) ) {
-			$t_ids_to_seach[(int)$t_id] = (int)$t_id;
+			$t_ids_to_search[(int)$t_id] = (int)$t_id;
 		}
 	}
-	if( empty( $t_ids_to_seach ) ) {
+	if( empty( $t_ids_to_search ) ) {
 		return;
 	}
 
 	db_param_push();
 	$t_sql_in_params = array();
 	$t_params = array();
-	foreach( $t_ids_to_seach as $t_id ) {
+	foreach( $t_ids_to_search as $t_id ) {
 		$t_sql_in_params[] = db_param();
 		$t_params[] = $t_id;
 	}
@@ -96,10 +98,10 @@ function tag_cache_rows( array $p_tag_ids ) {
 	while( $t_row = db_fetch_array( $t_result ) ) {
 		$c_id = (int)$t_row['id'];
 		$g_cache_tags[$c_id] = $t_row;
-		unset( $t_ids_to_seach[$c_id] );
+		unset( $t_ids_to_search[$c_id] );
 	}
 	# mark the non existent ids
-	foreach( $t_ids_to_seach as $t_id ) {
+	foreach( $t_ids_to_search as $t_id ) {
 		$g_cache_tags[$t_id] = false;
 	}
 }
@@ -114,20 +116,21 @@ function tag_cache_rows( array $p_tag_ids ) {
 function tag_cache_bug_tag_rows( array $p_bug_ids ) {
 	global $g_cache_bug_tags;
 
-	$t_ids_to_seach = array();
+	$t_ids_to_search = array();
 	foreach( $p_bug_ids as $t_id ) {
 		if( !isset( $g_cache_bug_tags[(int)$t_id]) ) {
-			$t_ids_to_seach[] = (int)$t_id;
+			$t_ids_to_search[] = (int)$t_id;
 		}
 	}
-	if( empty( $t_ids_to_seach ) ) {
+
+	if( empty( $t_ids_to_search ) ) {
 		return;
 	}
 
 	db_param_push();
 	$t_sql_in_params = array();
 	$t_params = array();
-	foreach( $t_ids_to_seach as $t_id ) {
+	foreach( $t_ids_to_search as $t_id ) {
 		$t_sql_in_params[] = db_param();
 		$t_params[] = $t_id;
 	}
@@ -189,8 +192,10 @@ function tag_exists( $p_tag_id ) {
  */
 function tag_ensure_exists( $p_tag_id ) {
 	if( !tag_exists( $p_tag_id ) ) {
-		error_parameters( $p_tag_id );
-		trigger_error( ERROR_TAG_NOT_FOUND, ERROR );
+		throw new ClientException(
+			sprintf( "Tag '%d' does not exist", $p_tag_id ),
+			ERROR_TAG_NOT_FOUND,
+			array( $p_tag_id ) );
 	}
 }
 
@@ -323,7 +328,7 @@ function tag_attach_many( $p_bug_id, $p_tag_string, $p_tag_id = 0 ) {
 	access_ensure_bug_level( config_get( 'tag_attach_threshold' ), $p_bug_id );
 
 	$t_tags = tag_parse_string( $p_tag_string );
-	$t_can_create = access_has_global_level( config_get( 'tag_create_threshold' ) );
+	$t_can_create = tag_can_create();
 
 	$t_tags_create = array();
 	$t_tags_attach = array();
@@ -389,7 +394,7 @@ function tag_parse_filters( $p_string ) {
 		if( !is_blank( $t_name ) && tag_name_is_valid( $t_name, $t_matches, $t_prefix ) ) {
 			$t_tag_row = tag_get_by_name( $t_matches[1] );
 			if( $t_tag_row !== false ) {
-				$t_filter = utf8_substr( $t_name, 0, 1 );
+				$t_filter = mb_substr( $t_name, 0, 1 );
 
 				if( '+' == $t_filter ) {
 					$t_tag_row['filter'] = 1;
@@ -476,6 +481,20 @@ function tag_get( $p_tag_id ) {
 }
 
 /**
+ * Get tag name by id.
+ * @param integer $p_tag_id The tag ID to retrieve from the database.
+ * @return string tag name or empty string if not found.
+ */
+function tag_get_name( $p_tag_id ) {
+	$t_tag_row = tag_get( $p_tag_id );
+	if( $t_tag_row === false ) {
+		return '';
+	}
+
+	return $t_tag_row['name'];
+}
+
+/**
  * Return a tag row for the given name.
  * @param string $p_name The tag name to retrieve from the database.
  * @return array|boolean Tag row
@@ -513,6 +532,26 @@ function tag_get_field( $p_tag_id, $p_field_name ) {
 }
 
 /**
+ * Can the specified user create a tag?
+ *
+ * @param integer $p_user_id The id of the user to check access rights for.
+ * @return bool true: can create, false: otherwise.
+ */
+function tag_can_create( $p_user_id = null ) {
+	return access_has_global_level( config_get( 'tag_create_threshold' ), $p_user_id );
+}
+
+/**
+ * Ensure specified user can create tags.
+ *
+ * @param integer $p_user_id The id of the user to check access rights for.
+ * @return void
+ */
+function tag_ensure_can_create( $p_user_id = null ) {
+	access_ensure_global_level( config_get( 'tag_create_threshold' ), $p_user_id );
+}
+
+/**
  * Create a tag with the given name, creator, and description.
  * Defaults to the currently logged in user, and a blank description.
  * @param string  $p_name        The tag name to create.
@@ -521,7 +560,7 @@ function tag_get_field( $p_tag_id, $p_field_name ) {
  * @return int Tag ID
  */
 function tag_create( $p_name, $p_user_id = null, $p_description = '' ) {
-	access_ensure_global_level( config_get( 'tag_create_threshold' ) );
+	tag_ensure_can_create( $p_user_id );
 
 	tag_ensure_name_is_valid( $p_name );
 	tag_ensure_unique( $p_name );
@@ -576,7 +615,7 @@ function tag_update( $p_tag_id, $p_name, $p_user_id, $p_description ) {
 	tag_ensure_name_is_valid( $p_name );
 
 	$t_rename = false;
-	if( utf8_strtolower( $p_name ) != utf8_strtolower( $t_tag_name ) ) {
+	if( mb_strtolower( $p_name ) != mb_strtolower( $t_tag_name ) ) {
 		tag_ensure_unique( $p_name );
 		$t_rename = true;
 	}
@@ -831,7 +870,12 @@ function tag_bug_detach( $p_tag_id, $p_bug_id, $p_add_history = true, $p_user_id
 		$t_detach_level = config_get( 'tag_detach_threshold' );
 	}
 
-	access_ensure_bug_level( $t_detach_level, $p_bug_id, $t_user_id );
+	if( !access_has_bug_level( $t_detach_level, $p_bug_id, $t_user_id ) ) {
+		throw new ClientException(
+			sprintf( "Access denied to detach '%s'", $t_tag_row['name'] ),
+			ERROR_ACCESS_DENIED
+		);
+	}
 
 	db_param_push();
 	$t_query = 'DELETE FROM {bug_tag} WHERE tag_id=' . db_param() . ' AND bug_id=' . db_param();
@@ -988,34 +1032,18 @@ function tag_stats_related( $p_tag_id, $p_limit = 5 ) {
 	);
 	$t_filter = filter_ensure_valid_filter( $t_filter );
 
-	# Note: filter_get_bug_rows_query_clauses() calls db_param_push();
-	$t_query_clauses = filter_get_bug_rows_query_clauses( $t_filter, null, null, null );
-	# if the query can't be formed, there are no results
-	if( empty( $t_query_clauses ) ) {
-		# reset the db_param stack that was initialized by "filter_get_bug_rows_query_clauses()"
-		db_param_pop();
-		return array();
-	}
-	$t_select_string = 'SELECT {bug}.id ';
-	$t_from_string = ' FROM ' . implode( ', ', $t_query_clauses['from'] );
-	$t_join_string = count( $t_query_clauses['join'] ) > 0 ? implode( ' ', $t_query_clauses['join'] ) : ' ';
-	$t_where_string = ' WHERE '. implode( ' AND ', $t_query_clauses['project_where'] );
-	if( count( $t_query_clauses['where'] ) > 0 ) {
-		$t_where_string .= ' AND ( ' . implode( $t_query_clauses['operator'], $t_query_clauses['where'] ) . ' ) ';
-	}
-	$t_filter_in = ' ( ' . $t_select_string . $t_from_string . $t_join_string . $t_where_string . ' )';
-	$t_params = $t_query_clauses['where_values'];
+	$t_filter_subquery = new BugFilterQuery( $t_filter, BugFilterQuery::QUERY_TYPE_IDS );
 
-	$t_query = 'SELECT tag_id, COUNT(1) AS tag_count FROM {bug_tag}'
-			. ' WHERE bug_id IN ' . $t_filter_in
-			. ' AND tag_id <> ' . db_param()
-			. ' GROUP BY tag_id ORDER BY COUNT(1) DESC';
-
-	$t_params[] = (int)$p_tag_id;
-	$t_result = db_query( $t_query, $t_params, $p_limit );
+	$t_sql = 'SELECT tag_id, COUNT(1) AS tag_count FROM {bug_tag}'
+			. ' WHERE bug_id IN :filter AND tag_id <> :tagid'
+			. ' GROUP BY tag_id ORDER BY tag_count DESC';
+	$t_query = new DbQuery( $t_sql );
+	$t_query->bind( 'filter', $t_filter_subquery );
+	$t_query->bind( 'tagid', (int)$p_tag_id );
+	$t_query->execute();
 
 	$t_tags = array();
-	while( $t_row = db_fetch_array( $t_result ) ) {
+	while( $t_row = $t_query->fetch() ) {
 		$t_tag_row = tag_get( $t_row['tag_id'] );
 		$t_tag_row['count'] = (int)$t_row['tag_count'];
 		$t_tags[] = $t_tag_row;
